+++ app/uk/gov/hmrc/cdsreimbursementclaimfrontend/controllers/overpaymentsscheduled/SelectDutyCodesController.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.cdsreimbursementclaimfrontend.controllers.claims
+package uk.gov.hmrc.cdsreimbursementclaimfrontend.controllers.overpaymentsscheduled
 
 import cats.data.OptionT
 import cats.implicits.catsSyntaxOptionId
@@ -68,7 +68,7 @@ class SelectDutyCodesController @Inject() (
       Future.successful(Redirect(claimRoutes.SelectDutyTypesController.showDutyTypes(JourneyBindable.Scheduled)))
 
     def start(dutyType: DutyType): Future[Result] =
-      Future(Redirect(claimRoutes.SelectDutyCodesController.showDutyCodes(dutyType)))
+      Future(Redirect(routes.SelectDutyCodesController.showDutyCodes(dutyType)))
 
     withAnswers[SelectedDutyTaxCodesReimbursementAnswer] { (_, maybeAnswer) =>
       maybeAnswer.flatMap(_.value.headOption).fold(selectDuties) { selectedDutyTaxCodesReimbursement =>
@@ -81,7 +81,7 @@ class SelectDutyCodesController @Inject() (
     implicit request =>
       withAnswers[SelectedDutyTaxCodesReimbursementAnswer] { (_, maybeAnswer) =>
         val maybeSelectedTaxCodes = maybeAnswer.map(_.getTaxCodes(dutyType))
-        val postAction: Call      = claimRoutes.SelectDutyCodesController.submitDutyCodes(dutyType)
+        val postAction: Call      = routes.SelectDutyCodesController.submitDutyCodes(dutyType)
         Ok(
           selectDutyCodesPage(
             dutyType,
@@ -95,7 +95,7 @@ class SelectDutyCodesController @Inject() (
   def submitDutyCodes(currentDuty: DutyType): Action[AnyContent] =
     authenticatedActionWithSessionData.async { implicit request =>
       withAnswers[SelectedDutyTaxCodesReimbursementAnswer] { (fillingOutClaim, maybeAnswer) =>
-        val postAction: Call = claimRoutes.SelectDutyCodesController.submitDutyCodes(currentDuty)
+        val postAction: Call = routes.SelectDutyCodesController.submitDutyCodes(currentDuty)
 
         def updateClaim(answer: SelectedDutyTaxCodesReimbursementAnswer) = {
           val claim = from(fillingOutClaim)(_.copy(selectedDutyTaxCodesReimbursementAnswer = answer.some))
@@ -116,8 +116,8 @@ class SelectDutyCodesController @Inject() (
                     _ =>
                       maybeAnswer
                         .flatMap(_.findNextSelectedDutyAfter(currentDuty))
-                        .fold(Redirect(claimRoutes.EnterScheduledClaimController.iterate()))(nextDuty =>
-                          Redirect(claimRoutes.SelectDutyCodesController.showDutyCodes(nextDuty))
+                        .fold(Redirect(routes.EnterScheduledClaimController.iterate()))(nextDuty =>
+                          Redirect(routes.SelectDutyCodesController.showDutyCodes(nextDuty))
                         )
                   )
                 )
