+++ app/uk/gov/hmrc/cdsreimbursementclaimfrontend/views/claims/check_scheduled_claim_summary.scala.html
@@ -19,7 +19,7 @@
 @import play.twirl.api.Html
 @import uk.gov.hmrc.cdsreimbursementclaimfrontend.config.ViewConfig
 @import uk.gov.hmrc.cdsreimbursementclaimfrontend.controllers.actions.RequestWithSessionData
-@import uk.gov.hmrc.cdsreimbursementclaimfrontend.controllers.claims.{routes => claimRoutes}
+@import uk.gov.hmrc.cdsreimbursementclaimfrontend.controllers.overpaymentsscheduled.{routes => overpaymentsScheduledRoutes}
 @import uk.gov.hmrc.cdsreimbursementclaimfrontend.models.answers.{SelectedDutyTaxCodesReimbursementAnswer, YesNo}
 @import uk.gov.hmrc.cdsreimbursementclaimfrontend.views.components.summary.DutyAndTaxCodeReimbursementSummary
 @import uk.gov.hmrc.cdsreimbursementclaimfrontend.views.components.summary.TaxCodeReimbursementSummary
@@ -42,7 +42,7 @@
 @key = @{"check-claim-summary"}
 @title = @{messages(s"$key.scheduled.title")}
 
-@postAction = @{claimRoutes.CheckScheduledClaimController.submitReimbursements()}
+@postAction = @{overpaymentsScheduledRoutes.CheckScheduledClaimController.submitReimbursements()}
 
 @hasErrors = @{form.hasErrors || form.hasGlobalErrors}
